Refactor sidecar worker entrypoint
